backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0
python-multipart==0.0.6
python-dotenv==1.0.0
yt-dlp==2023.11.16
opencv-python==4.8.1.78
numpy==1.26.2
redis==5.0.1
pillow==10.1.0
email-validator==2.1.0
httpx==0.28.1
slowapi==0.1.9
